refactor: move public header to ui components
