+++ project/app/(auth)/sign-in/[[...sign-in]]/page.tsx
@@ -1,5 +1,5 @@
-import { Header } from "@/components/header";
 import SignInForm from "@/components/sections/sign-in/sign-in-form";
+import { Header } from "@/components/ui/header";
 import { Fragment } from "react";
 
 export default function SignInPage() {
