+++ project/app/(auth)/sign-up/[[...sign-up]]/page.tsx
@@ -1,5 +1,5 @@
-import { Header } from "@/components/header";
 import SignUpForm from "@/components/sections/sign-up/sign-up-form";
+import { Header } from "@/components/ui/header";
 import { Fragment } from "react";
 
 export default function SignUpPage() {
