+++ project/app/page.tsx
@@ -1,5 +1,5 @@
-import { Header } from "@/components/header";
 import Hero from "@/components/sections/home/hero";
+import { Header } from "@/components/ui/header";
 import { Fragment } from "react";
 
 export default function HomePage() {
